child_process: use ERR_INVALID_SIGNATURE
Throw ERR_INVALID_SIGNATURE in execFile() and fork() for invalid
signatures, and include a message test for those errors. Also
refactors test-child-process-spawn-typeerror.js and rename it
to test-child-process-argument-errors.js
