+++ lib/child_process.js
@@ -65,9 +65,10 @@ exports.fork = function(modulePath /*, args, options*/) {
 
   if (pos < arguments.length && arguments[pos] != null) {
     if (typeof arguments[pos] !== 'object') {
-      throw new errors.TypeError('ERR_INVALID_ARG_VALUE',
-                                 `arguments[${pos}]`,
-                                 arguments[pos]);
+      throw new errors.TypeError(
+        'ERR_INVALID_SIGNATURE',
+        'child_process.fork', '(modulePath[, args][, options])',
+        [...arguments]);
     }
 
     options = util._extend({}, arguments[pos++]);
@@ -200,7 +201,10 @@ exports.execFile = function(file /*, args, options, callback*/) {
   }
 
   if (!callback && pos < arguments.length && arguments[pos] != null) {
-    throw new errors.TypeError('ERR_INVALID_ARG_VALUE', 'args', arguments[pos]);
+    throw new errors.TypeError(
+      'ERR_INVALID_SIGNATURE',
+      'child_process.execFile', '(file[, args][, options][, callback])',
+      [...arguments]);
   }
 
   // Validate the timeout, if present.
